+++ tilings/strategies/__init__.py
@@ -43,6 +43,7 @@ from .sliding import SlidingFactory
 from .symmetry import SymmetriesFactory
 from .verification import (
     BasicVerificationStrategy,
+    ComponentVerificationStrategy,
     DatabaseVerificationStrategy,
     ElementaryVerificationStrategy,
     InsertionEncodingVerificationStrategy,
@@ -99,6 +100,7 @@ __all__ = [
     "SymmetriesFactory",
     # Verification
     "BasicVerificationStrategy",
+    "ComponentVerificationStrategy",
     "DatabaseVerificationStrategy",
     "ElementaryVerificationStrategy",
     "LocallyFactorableVerificationStrategy",